feat(audio): make the WASM audio-processing backend pluggable per provider
TL;DR: extract BBBA's WASM processing into its own provider module, add
a config knob to pick the active provider, and let a provider force
browser-constraint overrides that always win over admin config.

audio-processor/service.js implements BBBA's WASM processing directly,
with no separation between the generic WASM/AudioWorklet orchestration
and BBBA's own fetch/worklet/port-message protocol. Adding a second
noise-suppression provider (WorkAdventure/DTLN, upcoming) on top of that
would mean branching provider-specific code inline in the same file, and
there is no config knob for an admin to pick one over the other.

Move BBBA's implementation into its own provider module conforming to a
small contract (isSupported/loadFiles/createProcessorStream/
forcedMicrophoneConstraints), and factor the generic browser-capability
probe out to a shared helper so a second provider can reuse it without
duplicating the WebAssembly.validate byte checks. service.js resolves the
active provider through a PROVIDERS registry and
getActiveProviderId()/getActiveProvider(), reading a new
media.audio.audioWasmProcessing.provider setting (admin-facing in
settings.yml, plus the JS/TS defaults) that accepts 'bbba' or
'workadventureDtln' and defaults to 'bbba', so existing deployments see no
behavior change.

doGUM() also hardcodes a single default for the browser-level constraints
applied alongside WASM processing, tuned for BBBA. A provider can instead
declare forcedMicrophoneConstraints - values that must win over whatever
an admin configures in audioWasmProcessing.constraints, because running
the browser's native filter alongside that provider's own processing
would hurt rather than help. That can't be expressed as just another
overridable default: settings.yml ships audioWasmProcessing.constraints
non-empty (noiseSuppression: true) by default, so a merely-overridable
value would silently lose to it for every deployment that hasn't
customized that block. doGUM() now merges forcedMicrophoneConstraints
last, so nothing in config can override it; empty for BBBA today.

Also fix doGUM()'s getUserMedia retry ladder: two fallback branches reset
to a bare { audio: true } on OverconstrainedError/NotFoundError/
NotReadableError, discarding every constraint already resolved (defaults,
admin config, and now forcedMicrophoneConstraints). This was harmless for
BBBA by coincidence - the browser's own default for an unset constraint is
true in Firefox/Chrome/Safari alike - but would silently defeat a
provider's forced noiseSuppression: false, and the fallback is reachable
from the main audio-join path, not just the settings preview. Add
withoutDeviceIdConstraint() so the fallback only gives up on deviceId, not
every other resolved constraint. Rename isBBBAWasmSupported to
isAdvancedProcessingSupported, since it's a generic capability check with
zero external callers, not BBBA-specific.
