+++ bigbluebutton-html5/private/config/settings.yml
@@ -905,12 +905,25 @@ public:
       # by closing the audio modal while audio is being negotiated.
       # If false, the audio modal will be locked until the process finishes.
       allowAudioJoinCancel: true
-      # audioWasmProcessing: audio input processing through WASM/BBBA
-      #   enabled: whether BBBA should be exposed to users. If false, browser's
-      #            built-in audio processing will be used instead.
+      # audioWasmProcessing: audio input processing through WASM
+      #   enabled: whether advanced/WASM processing should be exposed to users.
+      #            If false, browser's built-in audio processing will be used instead.
+      #   provider: which WASM backend runs the processing - 'bbba' or
+      #             'workadventureDtln'. Defaults to 'bbba' if unset/unrecognized.
       #   constraints: browser-level audio constraints applied ON TOP of WASM processing.
+      #                Not every value here is guaranteed to reach the browser as-is:
+      #                a provider can force its own value for a given constraint,
+      #                overriding whatever is configured below, when running the
+      #                browser's native filter alongside its WASM processing would
+      #                hurt rather than help. A provider's forced value always wins.
+      #                Each provider declares its overrides (if any) via
+      #                forcedMicrophoneConstraints, in
+      #                imports/ui/components/audio/audio-processor/providers/:
+      #                  - bbba: none. Every constraint below is honored as configured.
+      #                    (providers/bbba.js)
       audioWasmProcessing:
         enabled: false
+        provider: bbba
         constraints:
           echoCancellation: true
           autoGainControl: true
